move cli main into `lib/`
